+++ src/php/Compiler/Application/GlobalEnvironmentManager.php
@@ -9,6 +9,7 @@ use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironmentInterface;
 use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironmentManagerInterface;
 use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironmentRegistry;
 use Phel\Compiler\Domain\Evaluator\RequireEvaluator;
+use Phel\Shared\TagResolver;
 
 /**
  * @internal
@@ -28,6 +29,7 @@ final class GlobalEnvironmentManager implements GlobalEnvironmentManagerInterfac
     public function reset(): void
     {
         GlobalEnvironmentRegistry::set(null);
+        TagResolver::setUseAliasResolver(null);
     }
 
     public function isInitialized(): bool